fix(splitButton): let ui-bootstrap handle isOpen
- fixes 'dropdown closes too soon' issue in Safari:
- Ui-bootstrap's behaviour of opening on focus did not work with
the split-control-button directive. The previous workaround
set isOpen = true in $element.focus() caused problems in Safari -
the event was triggered on mouse click as well as on
keyboard focus.  On mouse click, ui-bootstrap toggles isOpen so
the menu auto-closed after it opens.
- Instead this let ui-bootstrap manage isOpen but listen for
'enter' keyboard events on the button then sends a click event.
- add tabindex=0 to dropdown items so can tab through them
Closes #1978