GDCDEVOPS-496 assume deployment at /uat prefix
Update various configs and links to assume the portal is deployed
under a /uat prefix. Keep the auth API at the usual location since NIH
requires login to redirect to a specific URL.

Update hardcoded JS links in index.html to use %PUBLIC_URL%.

Access the API using relative URLs so we're not tied to a specific
domain.