GDCDEVOPS-494 simplify deployment at different paths
Use `GDC_BASE` to establish the base URL of the portal, and make
other path-related variables relative to that. Introduce new
`LEGACY_PORTAL` and `SUBMISSION_PORTAL` variables so that links to
the legacy and submission portal can also be relative. Make paths
relative by default so that API requests will go to whatever domaain
the portal itself is deployed it.

Make it easier to deploy the portal under new domains, or under new
subpaths wihin a domain.